rename: AddBillFunnel 경로 수정
