+++ client/src/hooks/useAddBillFunnel.ts
@@ -1,6 +1,6 @@
 import {useEffect, useState} from 'react';
 
-import {BillInfo} from '@pages/BillPage/AddBillFunnel';
+import {BillInfo} from '@pages/AddBillFunnel/AddBillFunnel';
 
 import useRequestGetCurrentMembers from './queries/member/useRequestGetCurrentMembers';
